splitted links in their owner repository
