+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/repository/ClaimedTaskRestRepository.java
@@ -9,6 +9,7 @@ package org.dspace.app.rest.repository;
 
 import java.io.IOException;
 import java.sql.SQLException;
+import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
 import javax.mail.MessagingException;
@@ -16,6 +17,7 @@ import javax.servlet.http.HttpServletRequest;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.log4j.Logger;
+import org.dspace.app.rest.DiscoverableEndpointsService;
 import org.dspace.app.rest.Parameter;
 import org.dspace.app.rest.SearchRestMethod;
 import org.dspace.app.rest.exception.RESTAuthorizationException;
@@ -42,10 +44,12 @@ import org.dspace.xmlworkflow.state.actions.WorkflowActionConfig;
 import org.dspace.xmlworkflow.storedcomponents.ClaimedTask;
 import org.dspace.xmlworkflow.storedcomponents.XmlWorkflowItem;
 import org.dspace.xmlworkflow.storedcomponents.service.ClaimedTaskService;
+import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
+import org.springframework.hateoas.Link;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Component;
 
@@ -56,7 +60,8 @@ import org.springframework.stereotype.Component;
  */
 
 @Component(PoolTaskRest.CATEGORY + "." + ClaimedTaskRest.NAME)
-public class ClaimedTaskRestRepository extends DSpaceRestRepository<ClaimedTaskRest, Integer> {
+public class ClaimedTaskRestRepository extends DSpaceRestRepository<ClaimedTaskRest, Integer>
+                                       implements InitializingBean {
 
     private static final Logger log = Logger.getLogger(ClaimedTaskRestRepository.class);
 
@@ -77,6 +82,9 @@ public class ClaimedTaskRestRepository extends DSpaceRestRepository<ClaimedTaskR
 
     @Autowired
     AuthorizeService authorizeService;
+
+    @Autowired
+    DiscoverableEndpointsService discoverableEndpointsService;
 
     @Override
     @PreAuthorize("hasPermission(#id, 'CLAIMEDTASK', 'READ')")
@@ -187,5 +195,12 @@ public class ClaimedTaskRestRepository extends DSpaceRestRepository<ClaimedTaskR
 
     public Page<ClaimedTaskRest> findAll(Context context, Pageable pageable) {
         throw new RepositoryMethodNotImplementedException(ClaimedTaskRest.NAME, "findAll");
+    }
+
+    @Override
+    public void afterPropertiesSet() throws Exception {
+        discoverableEndpointsService.register(this, Arrays.asList(
+                new Link("/api/" + ClaimedTaskRest.CATEGORY + "/" + ClaimedTaskRest.NAME + "/search",
+                        ClaimedTaskRest.NAME + "-search")));
     }
 }
