+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/repository/PoolTaskRestRepository.java
@@ -9,12 +9,14 @@ package org.dspace.app.rest.repository;
 
 import java.io.IOException;
 import java.sql.SQLException;
+import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
 import javax.mail.MessagingException;
 import javax.servlet.http.HttpServletRequest;
 
 import org.apache.log4j.Logger;
+import org.dspace.app.rest.DiscoverableEndpointsService;
 import org.dspace.app.rest.Parameter;
 import org.dspace.app.rest.SearchRestMethod;
 import org.dspace.app.rest.exception.RESTAuthorizationException;
@@ -36,10 +38,12 @@ import org.dspace.xmlworkflow.state.Workflow;
 import org.dspace.xmlworkflow.state.actions.WorkflowActionConfig;
 import org.dspace.xmlworkflow.storedcomponents.PoolTask;
 import org.dspace.xmlworkflow.storedcomponents.service.PoolTaskService;
+import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
+import org.springframework.hateoas.Link;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Component;
 
@@ -50,7 +54,8 @@ import org.springframework.stereotype.Component;
  */
 
 @Component(PoolTaskRest.CATEGORY + "." + PoolTaskRest.NAME)
-public class PoolTaskRestRepository extends DSpaceRestRepository<PoolTaskRest, Integer> {
+public class PoolTaskRestRepository extends DSpaceRestRepository<PoolTaskRest, Integer>
+                                    implements InitializingBean {
 
     private static final Logger log = Logger.getLogger(PoolTaskRestRepository.class);
 
@@ -71,6 +76,9 @@ public class PoolTaskRestRepository extends DSpaceRestRepository<PoolTaskRest, I
 
     @Autowired
     AuthorizeService authorizeService;
+
+    @Autowired
+    DiscoverableEndpointsService discoverableEndpointsService;
 
     @Override
     @PreAuthorize("hasPermission(#id, 'POOLTASK', 'READ')")
@@ -145,5 +153,12 @@ public class PoolTaskRestRepository extends DSpaceRestRepository<PoolTaskRest, I
     @Override
     public Page<PoolTaskRest> findAll(Context context, Pageable pageable) {
         throw new RuntimeException("Method not allowed!");
+    }
+
+    @Override
+    public void afterPropertiesSet() throws Exception {
+        discoverableEndpointsService.register(this, Arrays.asList(
+                new Link("/api/" + PoolTaskRest.CATEGORY + "/" + PoolTaskRest.NAME + "/search",
+                        PoolTaskRest.NAME + "-search")));
     }
 }
