+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/repository/ResourcePolicyRestRepository.java
@@ -9,11 +9,13 @@ package org.dspace.app.rest.repository;
 
 import java.io.IOException;
 import java.sql.SQLException;
+import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
 import javax.servlet.http.HttpServletRequest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.dspace.app.rest.DiscoverableEndpointsService;
 import org.dspace.app.rest.Parameter;
 import org.dspace.app.rest.SearchRestMethod;
 import org.dspace.app.rest.exception.MissingParameterException;
@@ -35,10 +37,12 @@ import org.dspace.eperson.EPerson;
 import org.dspace.eperson.Group;
 import org.dspace.eperson.service.EPersonService;
 import org.dspace.eperson.service.GroupService;
+import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
+import org.springframework.hateoas.Link;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Component;
 
@@ -48,7 +52,8 @@ import org.springframework.stereotype.Component;
  * @author Luigi Andrea Pascarelli (luigiandrea.pascarelli at 4science.it)
  */
 @Component(ResourcePolicyRest.CATEGORY + "." + ResourcePolicyRest.NAME)
-public class ResourcePolicyRestRepository extends DSpaceRestRepository<ResourcePolicyRest, Integer> {
+public class ResourcePolicyRestRepository extends DSpaceRestRepository<ResourcePolicyRest, Integer>
+                                          implements InitializingBean {
 
     @Autowired
     ResourcePolicyService resourcePolicyService;
@@ -67,6 +72,9 @@ public class ResourcePolicyRestRepository extends DSpaceRestRepository<ResourceP
 
     @Autowired
     ResourcePatch<ResourcePolicy> resourcePatch;
+
+    @Autowired
+    DiscoverableEndpointsService discoverableEndpointsService;
 
     @Override
     @PreAuthorize("hasPermission(#id, 'resourcepolicy', 'READ')")
@@ -311,5 +319,12 @@ public class ResourcePolicyRestRepository extends DSpaceRestRepository<ResourceP
         }
         resourcePatch.patch(obtainContext(), resourcePolicy, patch.getOperations());
         resourcePolicyService.update(context, resourcePolicy);
+    }
+
+    @Override
+    public void afterPropertiesSet() throws Exception {
+        discoverableEndpointsService.register(this, Arrays.asList(
+                        new Link("/api/" + ResourcePolicyRest.CATEGORY + "/" + ResourcePolicyRest.NAME + "/search",
+                                ResourcePolicyRest.NAME + "-search")));
     }
 }
