+++ dspace-server-webapp/src/test/java/org/dspace/app/rest/EPersonRestRepositoryIT.java
@@ -1773,4 +1773,18 @@ public class EPersonRestRepositoryIT extends AbstractControllerIntegrationTest {
                             );
 
     }
+
+    @Test
+    public void discoverableNestedLinkTest() throws Exception {
+        String token = getAuthToken(eperson.getEmail(), password);
+        getClient(token).perform(get("/api"))
+                        .andExpect(status().isOk())
+                        .andExpect(jsonPath("$._links",Matchers.allOf(
+                                hasJsonPath("$.epersons.href",
+                                         is("http://localhost/api/eperson/epersons")),
+                                hasJsonPath("$.eperson-registration.href",
+                                         is("http://localhost/api/eperson/registrations"))
+                        )));
+    }
+
 }
