+++ dspace-server-webapp/src/test/java/org/dspace/app/rest/ResourcePolicyRestRepositoryIT.java
@@ -2404,4 +2404,17 @@ public class ResourcePolicyRestRepositoryIT extends AbstractControllerIntegratio
                 hasJsonPath("$.endDate", is(formatDate.format(endDate))),
                 hasJsonPath("$.description", nullValue()))));
     }
+
+    @Test
+    public void discoverableNestedLinkTest() throws Exception {
+        String token = getAuthToken(eperson.getEmail(), password);
+        getClient(token).perform(get("/api"))
+                        .andExpect(status().isOk())
+                        .andExpect(jsonPath("$._links",Matchers.allOf(
+                                hasJsonPath("$.resourcepolicies.href",
+                                         is("http://localhost/api/authz/resourcepolicies")),
+                                hasJsonPath("$.resourcepolicy-search.href",
+                                         is("http://localhost/api/authz/resourcepolicy/search"))
+                        )));
+    }
 }
